refactor: constants·utils 동명 파일과 barrel 정리
constants/와 utils/에 같은 파일명이 네 쌍 있어 import 경로의 디렉터리로만
어느 쪽인지 구분되던 상태였다. Url.ts는 ExternalLink.ts로 바뀌며 해소됐고,
남은 세 쌍은 설정값을 담은 constants 쪽 이름을 역할이 드러나게 바꾼다.

- constants/Error.ts → MemoError.ts
- constants/Sentry.ts → SentryDsn.ts
- constants/Supabase.ts → SupabaseConfig.ts
- utils/highlightService.test.ts → utils/Supabase.test.ts.
  이름은 utils/supabase/highlightService.ts를 가리키는데 실제로는
  utils/Supabase.ts를 테스트하고 있어 대상과 이름을 일치시킨다
- utils/index.ts에 memoSearchFilter 추가 (barrel에서 누락돼 있었다)
