+++ packages/shared/src/constants/index.ts
@@ -1,15 +1,10 @@
 export * from "./Analytics";
-export * from "./Cookie";
-export * from "./Error";
-export * from "./Extension";
+export * from "./ChromeExtension";
+export * from "./ExternalLink";
 export * from "./Highlight";
-export * from "./Language";
-export * from "./MotionVariants";
-export * from "./OAuth";
+export * from "./MemoError";
 export * from "./OpenAI";
 export * from "./Path";
-export * from "./Prompt";
 export * from "./QueryKey";
-export * from "./Sentry";
-export * from "./Supabase";
-export * from "./Url";
+export * from "./SentryDsn";
+export * from "./SupabaseConfig";
